+++ internal/auth/error.go
