+++ internal/go2lua/export.go
