+++ regclient/manifest/docker1.go
@@ -8,7 +8,7 @@ import (
 	dockerSchema1 "github.com/docker/distribution/manifest/schema1"
 	digest "github.com/opencontainers/go-digest"
 	ociv1 "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/regclient/regclient/pkg/wraperr"
+	"github.com/regclient/regclient/internal/wraperr"
 )
 
 const (
