+++ regclient/manifest/manifest.go
@@ -12,7 +12,7 @@ import (
 	dockerSchema2 "github.com/docker/distribution/manifest/schema2"
 	digest "github.com/opencontainers/go-digest"
 	ociv1 "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/regclient/regclient/pkg/wraperr"
+	"github.com/regclient/regclient/internal/wraperr"
 	"github.com/regclient/regclient/regclient/types"
 )
