+++ regclient/manifest/unknown.go
@@ -5,7 +5,7 @@ import (
 
 	digest "github.com/opencontainers/go-digest"
 	ociv1 "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/regclient/regclient/pkg/wraperr"
+	"github.com/regclient/regclient/internal/wraperr"
 )
 
 type unknown struct {
